Extracted  main source from the command line support.
